# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Eko CP",
        "description": "Dokumentasi API untuk data Fase, Mapel, dan CP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/fase/{faseId}/mapel": {
            "get": {
                "tags": [
                    "Mapel"
                ],
                "summary": "Ambil semua mapel berdasarkan ID fase",
                "operationId": "82e7aafe8bdc8681a36ed2ea7e7824fc",
                "parameters": [
                    {
                        "name": "faseId",
                        "in": "path",
                        "description": "ID Fase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data mapel berhasil diambil"
                    }
                }
            }
        },
        "/api/fase/{faseId}/mapel/{mapelId}/metode": {
            "get": {
                "tags": [
                    "Metode Pembelajaran"
                ],
                "summary": "Ambil metode pembelajaran berdasarkan fase dan mapel",
                "operationId": "94bd0bb169d4818a0d68aa9a5dd5cba8",
                "parameters": [
                    {
                        "name": "faseId",
                        "in": "path",
                        "description": "ID Fase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mapelId",
                        "in": "path",
                        "description": "ID Mapel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metode pembelajaran berhasil diambil"
                    }
                }
            }
        },
        "/api/fase/{faseId}/mapel/{mapelId}/mp/{metode}": {
            "get": {
                "tags": [
                    "Capaian Pembelajaran"
                ],
                "summary": "Ambil data CP berdasarkan fase, mapel, dan metode pembelajaran",
                "operationId": "22ed6799b0362df62e95399c7c8c063d",
                "parameters": [
                    {
                        "name": "faseId",
                        "in": "path",
                        "description": "ID Fase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mapelId",
                        "in": "path",
                        "description": "ID Mapel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "metode",
                        "in": "path",
                        "description": "Metode Pembelajaran",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data CP berhasil diambil"
                    }
                }
            }
        },
        "/api/cp/{id}": {
            "get": {
                "tags": [
                    "Capaian Pembelajaran"
                ],
                "summary": "Lihat detail CP berdasarkan ID",
                "operationId": "4f9a199d3da3942b031fc7c8d8c8c4b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID CP",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail CP berhasil diambil"
                    },
                    "404": {
                        "description": "CP tidak ditemukan"
                    }
                }
            }
        },
...
        "/api/kaidah/tema": {
            "get": {
                "tags": [
                    "Kaidah"
                ],
                "summary": "Ambil semua tema untuk kaidah",
                "operationId": "9f879d945d42fe137f54b171e9ae483b",
                "responses": {
                    "200": {
                        "description": "Daftar Tema berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "nama_tema": {
                                                        "type": "string",
                                                        "example": "Etika Lingkungan"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of Tema retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tema tidak ditemukan"
                    }
                }
            }
        },
        "/api/kaidah/tema/{tema_id}/kaidah": {
            "get": {
                "tags": [
                    "Kaidah"
                ],
                "summary": "Ambil semua kaidah berdasarkan tema ID",
                "operationId": "9d41aeb7f80ccd8093ff33cd6ef13f5a",
                "parameters": [
                    {
                        "name": "tema_id",
                        "in": "path",
                        "description": "ID Tema untuk kaidah",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar Kaidah berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "judul": {
                                                        "type": "string",
                                                        "example": "Kaidah menjaga lingkungan"
                                                    },
                                                    "isi": {
                                                        "type": "string",
                                                        "example": "Menjaga kebersihan merupakan bagian dari iman."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of Kaidah retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kaidah tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}